requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
altair=4.2.0=pypi_0
anyio=3.5.0=pypi_0
appnope=0.1.3=pyhd8ed1ab_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
async-generator=1.10=pypi_0
attrs=21.4.0=pypi_0
babel=2.10.1=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports-zoneinfo=0.2.1=pypi_0
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pypi_0
bleach=5.0.0=pypi_0
blinker=1.4=pypi_0
bqplot=0.12.33=pypi_0
branca=0.5.0=pypi_0
ca-certificates=2021.10.8=h033912b_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py37hf985489_2
cffi=1.15.0=pypi_0
charset-normalizer=2.0.12=pypi_0
chart-studio=1.1.0=pypi_0
click=8.0.4=pypi_0
colorlover=0.3.0=pypi_0
colour=0.1.5=pypi_0
cryptography=36.0.2=pypi_0
cufflinks=0.17.3=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.5.1=py37hd8d24ac_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
earthengine-api=0.1.306=pypi_0
ee-extra=0.0.13=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
fastjsonschema=2.15.3=pypi_0
ffmpeg-python=0.2.0=pypi_0
filelock=3.6.0=pypi_0
folium=0.12.1.post1=pypi_0
fonttools=4.33.2=pypi_0
future=0.18.2=pypi_0
gdown=4.4.0=pypi_0
geeadd=0.5.5=pypi_0
geemap=0.13.2=pypi_0
geocoder=1.38.1=pypi_0
geographiclib=1.52=pypi_0
geojson=2.5.0=pypi_0
geopy=2.2.0=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
google-api-core=2.7.2=pypi_0
google-api-python-client=1.12.11=pypi_0
google-auth=2.6.6=pypi_0
google-auth-httplib2=0.1.0=pypi_0
google-cloud-core=2.3.0=pypi_0
google-cloud-storage=2.3.0=pypi_0
google-crc32c=1.3.0=pypi_0
google-resumable-media=2.3.2=pypi_0
googleapis-common-protos=1.56.0=pypi_0
h11=0.13.0=pypi_0
httplib2=0.20.4=pypi_0
httplib2shim=0.0.3=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.11.3=pypi_0
importlib-resources=5.7.1=pypi_0
ipyevents=2.0.1=pypi_0
ipyfilechooser=0.6.0=pypi_0
ipykernel=6.12.1=py37h0a7177a_0
ipyleaflet=0.16.0=pypi_0
ipython=7.32.0=py37hf985489_0
ipython-genutils=0.2.0=pypi_0
ipytree=0.2.1=pypi_0
ipywidgets=7.7.0=pypi_0
jedi=0.18.1=py37hf985489_1
jinja2=3.1.1=pypi_0
json5=0.9.6=pypi_0
jsonschema=4.4.0=pypi_0
jupyter-server=1.16.0=pypi_0
jupyter_client=7.2.2=pyhd8ed1ab_1
jupyter_core=4.9.2=py37hf985489_0
jupyterlab=3.3.4=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.13.0=pypi_0
jupyterlab-widgets=1.1.0=pypi_0
kiwisolver=1.4.2=pypi_0
libcxx=13.0.1=hc203e6f_0
libffi=3.4.2=h0d85af4_5
libsodium=1.0.18=hbcb3906_1
libzlib=1.2.11=h6c3fc93_1014
logzero=1.7.0=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=pypi_0
nbclassic=0.3.7=pypi_0
nbclient=0.6.0=pypi_0
nbconvert=6.5.0=pypi_0
nbformat=5.3.0=pypi_0
ncurses=6.3=h96cf925_1
nest-asyncio=1.5.5=pyhd8ed1ab_0
notebook=6.4.11=pypi_0
notebook-shim=0.1.0=pypi_0
numpy=1.21.5=pypi_0
opencv-python=4.5.5.64=pypi_0
openssl=3.0.2=h6c3fc93_1
outcome=1.1.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.5=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.1.0=pypi_0
pip=22.0.4=pyhd8ed1ab_0
plotly=5.7.0=pypi_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.29=pyha770c72_0
protobuf=3.20.1=pypi_0
psutil=5.9.0=py37h69ee0a8_1
ptyprocess=0.7.0=pyhd3deb0d_0
pyarrow=7.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pycrs=1.0.2=pypi_0
pydeck=0.7.1=pypi_0
pygments=2.11.2=pyhd8ed1ab_0
pympler=1.0.1=pypi_0
pyopenssl=22.0.0=pypi_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyrsistent=0.18.1=pypi_0
pyshp=2.2.0=pypi_0
pysocks=1.7.1=pypi_0
python=3.7.12=hf3644f1_100_cpython
python-box=6.0.2=pypi_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytz=2022.1=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pyzmq=22.3.0=py37h8f778e5_1
ratelim=0.1.6=pypi_0
readline=8.1=h05e3726_0
requests=2.27.1=pypi_0
retrying=1.3.3=pypi_0
rsa=4.8=pypi_0
sankee=0.0.7=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.11.2=pypi_0
selenium=4.1.3=pypi_0
semver=2.13.0=pypi_0
send2trash=1.8.0=pypi_0
setuptools=62.0.0=py37hf985489_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.0=pypi_0
sniffio=1.2.0=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.37.1=hb516253_0
streamlit=1.8.1=pypi_0
tenacity=8.0.1=pypi_0
terminado=0.13.3=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.12=h5dbffcc_0
toml=0.10.2=pypi_0
toolz=0.11.2=pypi_0
tornado=6.1=py37h69ee0a8_3
tqdm=4.64.0=pypi_0
traitlets=5.1.1=pyhd8ed1ab_0
traittypes=0.2.1=pypi_0
trio=0.20.0=pypi_0
trio-websocket=0.9.2=pypi_0
typing-extensions=4.1.1=pypi_0
tzdata=2022.1=pypi_0
tzlocal=4.2=pypi_0
uritemplate=3.0.1=pypi_0
urllib3=1.26.9=pypi_0
validators=0.18.2=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=pypi_0
websocket-client=1.3.2=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
whitebox=2.1.2=pypi_0
whiteboxgui=0.7.0=pypi_0
widgetsnbextension=3.6.0=pypi_0
wsproto=1.1.0=pypi_0
xyzservices=2022.4.0=pypi_0
xz=5.2.5=haf1e3a3_1
zeromq=4.3.4=he49afe7_1
zipp=3.8.0=pypi_0
zlib=1.2.11=h6c3fc93_1014
